Allow per-question sorting by marks in tests table
