+++ src/app/components/elements/quiz/QuizProgressCommon.tsx
@@ -1,18 +1,9 @@
-import React, {useContext, useState} from "react";
-import {closeActiveModal, openActiveModal, useAppDispatch, useReturnQuizToStudentMutation} from "../../../state";
+import React, {ComponentProps, useContext, useLayoutEffect, useMemo, useRef, useState} from "react";
 import {Button} from "reactstrap";
-import {
-    ContentBaseDTO,
-    IsaacQuizDTO,
-    IsaacQuizSectionDTO,
-    Mark,
-    QuizAssignmentDTO,
-    QuizUserFeedbackDTO
-} from "../../../../IsaacApiTypes";
-import {AssignmentProgressPageSettingsContext} from "../../../../IsaacAppTypes";
-import {isDefined, isQuestion, PATHS, QUIZ_VIEW_STUDENT_ANSWERS_RELEASE_TIMESTAMP, siteSpecific} from "../../../services";
-import {IsaacSpinner} from "../../handlers/IsaacSpinner";
+import {AppAssignmentProgress, AssignmentProgressPageSettingsContext} from "../../../../IsaacAppTypes";
+import {siteSpecific} from "../../../services";
 import {Link} from "react-router-dom";
+import orderBy from "lodash/orderBy";
 
 export const ICON = siteSpecific(
     {
@@ -27,53 +18,7 @@ export const ICON = siteSpecific(
     }
 );
 
-interface ResultsTableProps {
-    assignment: QuizAssignmentDTO;
-    userFeedback: QuizUserFeedbackDTO[] | undefined;
-}
-
-interface ResultRowProps {
-    row: QuizUserFeedbackDTO;
-    assignment: QuizAssignmentDTO;
-}
-
-export function questionsInSection(section?: IsaacQuizSectionDTO) {
-    return section?.children?.filter(isQuestion) || [];
-}
-
-export function questionsInQuiz(quiz?: IsaacQuizDTO) {
-    const questions: ContentBaseDTO[] = [];
-    quiz?.children?.forEach(
-        section => {
-            questions.push(...questionsInSection(section));
-        }
-    );
-    return questions;
-}
-
 export const passMark = 0.75;
-
-export function markQuestionClasses(row: QuizUserFeedbackDTO, mark: Mark | undefined, totalOrUndefined: number | undefined) {
-    if (!row.user?.authorisedFullAccess) {
-        return "revoked";
-    }
-
-    const correct = mark?.correct as number;
-    const incorrect = mark?.incorrect as number;
-    const total = totalOrUndefined as number;
-
-    if (correct === total) {
-        return "completed";
-    } else if ((correct / total) >= passMark) {
-        return "passed";
-    } else if ((incorrect / total) > (1 - passMark)) {
-        return "failed";
-    } else if (correct > 0 || incorrect > 0) {
-        return "in-progress";
-    } else {
-        return "not-attempted";
-    }
-}
 
 export function formatMark(numerator: number, denominator: number, formatAsPercentage: boolean) {
     let result;
@@ -85,169 +30,202 @@ export function formatMark(numerator: number, denominator: number, formatAsPerce
     return result;
 }
 
-export function ResultRow({row, assignment}: ResultRowProps) {
-    const dispatch = useAppDispatch();
-    const pageSettings = useContext(AssignmentProgressPageSettingsContext);
-
-    const [dropdownOpen, setDropdownOpen] = useState(false);
-    const toggle = () => setDropdownOpen(prevState => !prevState);
-
-    const [returnQuizToStudent, {isLoading: returningQuizToStudent}] = useReturnQuizToStudentMutation();
-    const returnToStudent = () => {
-        const confirm = () => {
-            returnQuizToStudent({quizAssignmentId: assignment.id as number, userId: row.user?.id as number})
-                .then(() => dispatch(closeActiveModal()));
-        };
-        dispatch(openActiveModal({
-            closeAction: () => dispatch(closeActiveModal()),
-            title: "Allow another attempt?",
-            body: "This will allow the student to attempt the test again.",
-            buttons: [
-                <Button key={1} color="primary" outline target="_blank" onClick={() => dispatch(closeActiveModal())}>
-                    Cancel
-                </Button>,
-                <Button key={0} color="primary" target="_blank" onClick={confirm}>
-                    Confirm
-                </Button>,
-        ]
-        }));
-    };
-
-    const quiz = assignment?.quiz;
-    const sections: IsaacQuizSectionDTO[] = quiz?.children || [];
-
-    let message;
-    if (!row.user?.authorisedFullAccess) {
-        message = "Not sharing";
-    } else if (!row.feedback?.complete) {
-        message = "Not completed";
-    }
-    const authorisedAccessAndComplete = message === undefined;
-
-    const viewAnswersLink = `/test/attempt/feedback/${assignment?.id}/${row.user?.id}`;
-    const canViewQuizAnswers = isDefined(assignment?.creationDate) && (assignment?.creationDate?.valueOf() > QUIZ_VIEW_STUDENT_ANSWERS_RELEASE_TIMESTAMP);
-
-    return <tr className={`${row.user?.authorisedFullAccess ? "" : " not-authorised"}`} title={`${row.user?.givenName} ${row.user?.familyName}`}>
-        <td className="student-name">
-            {authorisedAccessAndComplete ?
-                <>
-                    <Button className="quiz-student-menu" color="link" onClick={toggle} disabled={returningQuizToStudent}>
-                        <div
-                            className="quiz-student-name"
-                        >
-                            {row.user?.givenName}
-                            <span className="d-none d-lg-inline"> {row.user?.familyName}</span>
-                        </div>
-                        <div className="quiz-student-menu-icon">
-                            {returningQuizToStudent ? <IsaacSpinner size="sm" /> : <img src="/assets/menu.svg" alt="Menu" />}
-                        </div>
-                    </Button>
-                    {!returningQuizToStudent && dropdownOpen && <>
-                        <div className="py-2 px-3">
-                            <Button size="sm" onClick={returnToStudent}>Allow another attempt</Button>
-                        </div>
-                        {canViewQuizAnswers && <div className="py-2 px-3">
-                            <Button size="sm" tag={Link} to={viewAnswersLink}>View answers</Button>
-                        </div>}
-                    </>}
-                </>
-            :   <>
-                    {row.user?.givenName}
-                    <span className="d-none d-lg-inline"> {row.user?.familyName}</span>
-                </>
-            }
-        </td>
-        {!authorisedAccessAndComplete && <td colSpan={sections.map(questionsInSection).flat().length + 1}>{message}</td>}
-        {authorisedAccessAndComplete && <>
-            {sections.map(section => {
-                const mark = row.feedback?.sectionMarks?.[section.id as string];
-                const outOf = quiz?.sectionTotals?.[section.id as string];
-                return questionsInSection(section).map(question => {
-                    const questionMark = row.feedback?.questionMarks?.[question.id as string] || {} as Mark;
-                    const icon =
-                        questionMark.correct === 1 ? ICON.correct :
-                        questionMark.incorrect === 1 ? ICON.incorrect :
-                        /* default */ ICON.notAttempted;
-                    return <td key={question.id} className={markQuestionClasses(row, mark, outOf)}>
-                        {icon}
-                    </td>;
-                }).flat();
-            })}
-            <th className="total-column">
-                {formatMark(row.feedback?.overallMark?.correct as number, quiz?.total as number, pageSettings.formatAsPercentage)}
-            </th>
-        </>}
-    </tr>;
+export interface QuestionType {
+    id?: string | undefined;
+    title?: string | undefined;
+    questionPartsTotal?: number | undefined;
 }
 
-export function ResultsTable({assignment, userFeedback}: ResultsTableProps) {
-    const sections: IsaacQuizSectionDTO[] = assignment.quiz?.children || [];
-    const quiz: IsaacQuizDTO | undefined = assignment.quiz;
+export interface ResultsTableProps<Q extends QuestionType> {
+    progress?: AppAssignmentProgress[];
+    questions: Q[];
+    header: JSX.Element;
+    getQuestionTitle: (question: Q) => JSX.Element;
+    assignmentAverages: number[];
+    assignmentTotalQuestionParts: number;
+    markClasses: (row: AppAssignmentProgress, assignmentTotalQuestionParts: number) => string;
+    markQuestionClasses: (row: AppAssignmentProgress, questionIndex: number) => string;
+    showQuestionPartBreakdown?: boolean;
+}
 
-    const fractionCorrect = (questionId: string) => {
-        if (!assignment.userFeedback) return [0, 0];
-        const marks = assignment.userFeedback.map(row => row.feedback?.questionMarks?.[questionId]);
-        const definedMarks = marks?.filter(isDefined);
-        if (!definedMarks || definedMarks.length === 0) return [0, 0];
-        const correct = definedMarks.reduce((p, c) => p + (c.correct ?? 0), 0);
-        const total = assignment.userFeedback.length * ((definedMarks[0].correct ?? 0) + (definedMarks[0].incorrect ?? 0) + (definedMarks[0].notAttempted ?? 0));
-        return [correct, total];
-    };
+export function ResultsTable<Q extends QuestionType>({progress, questions, header, getQuestionTitle, assignmentAverages, assignmentTotalQuestionParts, markClasses, markQuestionClasses, showQuestionPartBreakdown} : ResultsTableProps<Q>) {
+    const [selectedQuestionNumber, setSelectedQuestion] = useState(0);
+    const selectedQuestion: Q | undefined = questions[selectedQuestionNumber];
 
-    const quizAverages = sections.map(section => {
-        return section.children?.reduce((acc, question) => {
-            const [correct, total] = fractionCorrect(question.id ?? "");
-            return acc + (total > 0 ? correct / total : 0);
-        }, 0) ?? 0;
-    });
+    const pageSettings = useContext(AssignmentProgressPageSettingsContext);
 
-    const tableHeaderFooter = (isFooter = false) => {
-        return <tr>
-            {isFooter ? <th className="bg-white border-bottom student-name"/> : <></>}
-            {questionsInQuiz(quiz).map((question, index) => {
-                const [correct, total] = fractionCorrect(question.id ?? "");
-                return <th key={index + "" + question.id} className="border">
-                    {formatMark(correct, total, true)}
-                </th>;
-            }).flat()}
-            <th className="border-bottom total-column">
-                {(quizAverages && quiz?.total) ? formatMark(quizAverages.reduce((p, c) => p + c, 0), quiz?.total, true) : "N/A"}
-            </th>
-        </tr>;
-    };
+    type SortOrder = number | "name" | "totalQuestionPartPercentage" | "totalQuestionPercentage";
+    const [sortOrder, setSortOrder] = useState<SortOrder>("name");
+    const [reverseOrder, setReverseOrder] = useState(false);
+    
+    function isSelected(q: Q) {
+        return q == selectedQuestion ? "selected" : "";
+    }
 
-    return <>
-        <div className="progress-header">
-            {userFeedback
-            ? <>
-                <strong>{userFeedback.reduce((p, c) => p + (c.feedback?.complete ? 1 : 0), 0)}</strong> of <strong>{userFeedback.length}</strong>
-                {` students have completed the test `}
-            </>
-            : 'Preview '}
-            <Link to={`${PATHS.PREVIEW_TEST}/${assignment.quizId}/page/1`}>{assignment.quiz?.title}</Link>.
-        </div>
-        <div className={"progress-table-container mb-5"}>
-            <table className="progress-table border w-100">
+    function sortClasses(q: SortOrder) {
+        if (q == sortOrder) {
+            return "sorted" + (reverseOrder ? " reverse" : " forward");
+        } else {
+            return "";
+        }
+    }
+
+    function toggleSort(itemOrder: SortOrder) {
+        setSortOrder(itemOrder);
+        if (sortOrder == itemOrder) {
+            setReverseOrder(!reverseOrder);
+        } else {
+            setReverseOrder(false);
+        }
+    }
+
+    function sortItem(props: ComponentProps<"th"> & {itemOrder: SortOrder}) {
+        const {itemOrder, ...rest} = props;
+        const className = (props.className || "") + " " + sortClasses(itemOrder);
+        const clickToSelect = typeof itemOrder === "number" ? (() => setSelectedQuestion(itemOrder)) : undefined;
+        const sortArrows = (typeof itemOrder !== "number" || itemOrder === selectedQuestionNumber) ?
+            <button className="sort" onClick={() => {toggleSort(itemOrder);}}>
+                <span className="up" >▲</span>
+                <span className="down">▼</span>
+            </button>
+            : undefined;
+        return <th key={props.key} {...rest} className={className} onClick={clickToSelect}>{props.children}{sortArrows}</th>;
+    }
+
+    const semiSortedProgress = useMemo(() => orderBy(progress, (item) => {
+        return item.user.authorisedFullAccess && -item.notAttemptedPartResults.reduce(function(sum, increment) {return sum + increment;}, 0);
+    }, [reverseOrder ? "desc" : "asc"]), [progress, reverseOrder]);
+
+    const sortedProgress = useMemo(() => orderBy((semiSortedProgress), (item) => {
+            switch (sortOrder) {
+                case "name":
+                    return (item.user.familyName + ", " + item.user.givenName).toLowerCase();
+                case "totalQuestionPartPercentage":
+                    return -item.correctQuestionPartsCount;
+                case "totalQuestionPercentage":
+                    return -item.tickCount;
+                default:
+                    return -item.correctPartResults[sortOrder];
+            }
+        }, [reverseOrder ? "desc" : "asc"])
+    , [semiSortedProgress, reverseOrder, sortOrder]);
+
+    const tableHeaderFooter = <tr className="progress-table-header-footer">
+        {sortItem({key: "name", itemOrder: "name"})}
+        {questions.map((q, index) =>
+            sortItem({key: q.id, itemOrder: index, className: isSelected(q), children: `${assignmentAverages[index]}%`})
+        )}
+        {showQuestionPartBreakdown ? <>
+            {sortItem({key: "totalQuestionPartPercentage", itemOrder: "totalQuestionPartPercentage", className:"total-column left", children: "Total Parts"})}
+            {sortItem({key: "totalQuestionPercentage", itemOrder: "totalQuestionPercentage", className:"total-column right", children: "Total Qs"})}
+        </> : 
+        <>
+            {sortItem({key: "totalQuestionPartPercentage", itemOrder: "totalQuestionPartPercentage", className:"total-column", children: "Overall"})}
+        </>
+            
+        
+    }
+    </tr>;
+
+    const tableRef = useRef<HTMLTableElement>(null);
+
+    useLayoutEffect(() => {
+        const table = tableRef.current;
+        if (table) {
+            const parentElement = table.parentElement as HTMLElement;
+            const firstRow = (table.firstChild as HTMLTableSectionElement).firstChild as HTMLTableRowElement;
+            const questionTH = firstRow.children[selectedQuestionNumber + 1] as HTMLTableHeaderCellElement;
+
+            const offsetLeft = questionTH.offsetLeft;
+            const parentScrollLeft = parentElement.scrollLeft;
+            const parentLeft = parentScrollLeft + parentElement.offsetLeft + 130;
+            const width = questionTH.offsetWidth;
+
+            let newScrollLeft;
+
+            if (offsetLeft < parentLeft) {
+                newScrollLeft = parentScrollLeft + offsetLeft - parentLeft - width / 2;
+            } else {
+                const offsetRight = offsetLeft + width;
+                const parentRight = parentLeft + parentElement.offsetWidth - 260;
+                if (offsetRight > parentRight) {
+                    newScrollLeft = parentScrollLeft + offsetRight - parentRight + width / 2;
+                }
+            }
+            if (newScrollLeft != undefined) {
+                parentElement.scrollLeft = newScrollLeft;
+            }
+        }
+    }, [selectedQuestionNumber]);
+
+    return <div className="assignment-progress-progress">
+        {header}
+        {progress && progress.length > 0 && <>
+            <div className="progress-questions">
+                <Button color="tertiary" disabled={selectedQuestionNumber == 0}
+                    onClick={() => setSelectedQuestion(selectedQuestionNumber - 1)}>◄</Button>
+                <div>
+                    {getQuestionTitle(selectedQuestion)}
+                </div>
+                <Button color="tertiary" disabled={selectedQuestionNumber === questions.length - 1}
+                    onClick={() => setSelectedQuestion(selectedQuestionNumber + 1)}>►</Button>
+            </div>
+            <table ref={tableRef} className="progress-table w-100">
+                <thead>
+                    {tableHeaderFooter}
+                </thead>
                 <tbody>
-                    <tr className="bg-white">
-                        <th rowSpan={3} className="bg-white border-bottom student-name">&nbsp;</th>
-                            {sections.map(section => <th key={section.id} colSpan={questionsInSection(section).length} className="border font-weight-bold">
-                                {section.title}
-                            </th>)}
-                        <th rowSpan={2} className="border-bottom total-column">Overall</th>
-                    </tr>
-                    <tr className="bg-white">
-                        {questionsInQuiz(quiz).map((question, index) => <th key={question.id} className="border">
-                            {`Q${index + 1}`}
-                        </th>).flat()}
-                    </tr>
-                    {tableHeaderFooter()}
-                    {assignment.userFeedback?.map(row =>
-                        <ResultRow key={row.user?.id} row={row} assignment={assignment} />
-                    )}
-                    {tableHeaderFooter(true)}
+                    {sortedProgress.map((studentProgress) => {
+                        const fullAccess = studentProgress.user.authorisedFullAccess;
+                        return <tr key={studentProgress.user.id} className={`${markClasses(studentProgress, assignmentTotalQuestionParts)}${fullAccess ? "" : " not-authorised"}`} title={`${studentProgress.user.givenName + " " + studentProgress.user.familyName}`}>
+                            <th className="student-name">
+                                {fullAccess && pageSettings.isTeacher ?
+                                    <Link to={`/progress/${studentProgress.user.id}`} target="_blank">
+                                        {studentProgress.user.givenName}
+                                        <span
+                                            className="d-none d-lg-inline"> {studentProgress.user.familyName}</span>
+                                    </Link> :
+                                    <span>{studentProgress.user.givenName} {studentProgress.user.familyName}</span>
+                                }
+                            </th>
+                            {questions.map((q, index) =>
+                                <td key={q.id} className={isSelected(questions[index]) + " " + markQuestionClasses(studentProgress, index)} onClick={() => setSelectedQuestion(index)}>
+                                    {(assignmentTotalQuestionParts === questions.length) ?
+                                        studentProgress.correctPartResults[index] === 1 ? ICON.correct :
+                                        studentProgress.incorrectPartResults[index] === 1 ? ICON.incorrect :
+                                        /* default */ ICON.notAttempted
+                                    : 
+                                    fullAccess ? formatMark(studentProgress.correctPartResults[index],
+                                        questions[index].questionPartsTotal as number,
+                                        pageSettings.formatAsPercentage) : ""
+                                    }
+                                </td> 
+                            )}
+                            {showQuestionPartBreakdown ? <>
+                                <th className="total-column left" title={fullAccess ? undefined : "Not Sharing"}>
+                                    {fullAccess ? formatMark(studentProgress.correctQuestionPartsCount,
+                                        assignmentTotalQuestionParts,
+                                        pageSettings.formatAsPercentage) : ""}
+                                </th>
+                                <th className="total-column right" title={fullAccess ? undefined : "Not Sharing"}>
+                                    {fullAccess ? formatMark(studentProgress.tickCount,
+                                        questions.length,
+                                        pageSettings.formatAsPercentage) : ""}
+                                </th>
+                            </> : 
+                                <th className="total-column" title={fullAccess ? undefined : "Not Sharing"}>
+                                    {fullAccess ? formatMark(studentProgress.correctQuestionPartsCount,
+                                        assignmentTotalQuestionParts,
+                                        pageSettings.formatAsPercentage) : ""}
+                                </th>
+                            }
+                        </tr>;
+                    })}
                 </tbody>
+                <tfoot>
+                    {tableHeaderFooter}
+                </tfoot>
             </table>
-        </div>
-    </>;
+        </>}
+    </div>;
 }
